push-notifications/api/requirements.txt
Flask==3.0.3
marshmallow==3.22.0
flask-apispec==0.11.4
requests==2.32.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
Flask-Migrate==4.0.7
celery[redis]==5.4.0
pytz==2024.2
waitress==3.0.2
